refactor(drawer): extract LinkStateText into its own component
